+++ app/src/main/java/com/example/vehiclecontacting/Web/TalkController/TalkRepository.kt
@@ -1,10 +1,13 @@
 package com.example.vehiclecontacting.Web.TalkController
 
 import androidx.lifecycle.MutableLiveData
+import com.example.vehiclecontacting.Repository.ActivityCollector
 import com.example.vehiclecontacting.Repository.LogRepository
 import com.example.vehiclecontacting.Repository.StatusRepository
+import com.example.vehiclecontacting.Repository.WebRepository
 import com.example.vehiclecontacting.Web.UserController.UserRepository
 import com.example.vehiclecontacting.Web.WebService
+import com.example.vehiclecontacting.Widget.ToastView
 import java.lang.Exception
 import kotlin.concurrent.thread
 
@@ -22,6 +25,10 @@ object TalkRepository {
      * success：成功 （返回json talkList：聊天列表 counts：数据总量 pages：页面总数）
      */
     fun getTalk(cnt: Int, fromId: String, page: Int, toId: String): Int {
+        if (!WebRepository.isNetworkConnected()) {
+            ToastView(ActivityCollector.curActivity!!).show("网络错误")
+            return StatusRepository.CONNECT_WRONG
+        }
         val data = talkService.getTalk(cnt, fromId, page, toId)
         var msg = ""
         try {
@@ -46,6 +53,10 @@ object TalkRepository {
      * success：成功 （返回json talkMsgList：聊天列表 pages：页面总数 counts：数据量）
      */
     fun getTalkList(cnt: Int, id: String, page: Int): Int {
+        if (!WebRepository.isNetworkConnected()) {
+            ToastView(ActivityCollector.curActivity!!).show("网络错误")
+            return StatusRepository.CONNECT_WRONG
+        }
         val data = talkService.getTalkList(cnt, id, page)
         var msg = ""
         try {
@@ -71,6 +82,10 @@ object TalkRepository {
      * success：成功
      */
     fun deleteTalk(fromId: String, toId: String): Int {
+        if (!WebRepository.isNetworkConnected()) {
+            ToastView(ActivityCollector.curActivity!!).show("网络错误")
+            return StatusRepository.CONNECT_WRONG
+        }
         val data = talkService.deleteTalk(fromId, toId)
         var msg = ""
         try {
